% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Farsdata.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plots the accidents on a US state map}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{State number}

\item{years}{The year of concern}
}
\value{
\code{fars_map_state} will returns a plot of the accidents based on the \code{state.num}
 and \code{year} inputs. Returns an error if the state or year do not exist in the data set.
}
\description{
\code{fars_map_state} will plot the accidents on a map for a given state
  and year.
}
\examples{
\dontrun{
fars_map_state(42, 2013)
}

}
\seealso{
\code{\link{make_filename}} to understand how the file name is created
\code{\link{fars_read}} to understand how the file is read in
}
