% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Farsdata.R
\name{fars_read}
\alias{fars_read}
\title{Read in  data from the US National Highway Traffic Safety Administration's
Fatality Analysis Reporting System}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{csv file containing data}
}
\value{
\code{fars_read} will search within the specified path for the
  filename provided. If the file exists, it will be imported and returned as
  a data frame tbl.  If it does not exist an error message will be returned.
}
\description{
The function loads a CSV file defined by \code{fars_read} argument and returns
a tibble. If the path is incorrect the function will end with an error.
}
\examples{
\dontrun{
accident_2013 <- fars_read("./data/accident_2013.csv.bz2")
}
}
