% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Farsdata.R
\name{make_filename}
\alias{make_filename}
\title{Name the File}
\usage{
make_filename(year)
}
\arguments{
\item{year}{the year to add to the file name}
}
\value{
\code{make_file} will returns a character string in a format "accident_<year>.csv.bz2"
that can be used as a file name
}
\description{
\code{make_file} creates a name for the csv file for each year.
}
\examples{
\dontrun{
makefilename(2013)
}

}
